requirements.txt
selenium
nopo
ping3
loguru
python-dotenv